Fix parameter types for deflateInit2 to match expected int signature

+++ source/extensions/compression/gzip/compressor/zlib_compressor_impl.cc
@@ -36,8 +36,8 @@ void ZlibCompressorImpl::init(CompressionLevel comp_level, CompressionStrategy c
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wold-style-cast"
 #endif
-  const int result = deflateInit2(zstream_ptr_.get(), static_cast<int64_t>(comp_level), Z_DEFLATED,
-                                  window_bits, memory_level, static_cast<uint64_t>(comp_strategy));
+  const int result = deflateInit2(zstream_ptr_.get(), static_cast<int>(comp_level), Z_DEFLATED,
+                                  window_bits, memory_level, static_cast<int>(comp_strategy));
 #if defined(__GNUC__)
 #pragma GCC diagnostic pop
 #endif
